+++ micro/micro/doctype/micro_task/micro_task.py
@@ -1,8 +1,0 @@
-# SPDX-License-Identifier: AGPL-3.0-or-later
-# Copyright (C) 2026 Tonic
-
-from frappe.model.document import Document
-
-
-class MicroTask(Document):
-	pass